configured_plots.py: move get_pid_analysis_plots() into separate file
